requirements.txt
flask==2.3
flasgger==0.9.5
markdown==3.4.4
google-generativeai==0.8.4
pillow==10.4.0
ollama
pycryptodome==3.16.0
python-dotenv==1.0.0
pdfplumber==0.11.5
python-docx
